red: restructure, fix WithEnvironment
