+++ Decide/ObservableState/AtomicState.swift
@@ -14,6 +14,28 @@
 
 import Foundation
 
+
+/// AtomicState is a managed by ``ApplicationEnvironment`` container for ``Property`` and ``DefaultInstance`` definitions,
+/// its only requirement is to provide standalone `init()` so ``ApplicationEnvironment`` can instantiate it when necessary.
+/// You should never use instances of ``AtomicState`` directly, use ``Property`` or ``DefaultInstance`` instead.
+///
+/// **Usage:**
+/// ```swift
+/// final class TestState: AtomicState {
+///     // Declaration of a AtomicState property with a string value
+///     // that is "default-value" by default.
+///     @Property var name: String = "default-value"
+///
+///     // Declaration of the instance of a `NetworkingInterface` protocol
+///     // that is `Networking()` by default.
+///     @DefaultInstance var networking: NetworkingInterface = Networking()
+/// }
+/// ```
+@MainActor open class AtomicState {
+    required public init() {}
+}
+
+
 extension ApplicationEnvironment {
     subscript<S: AtomicState>(_ stateType: S.Type) -> S {
         let key = Key.atomic(ObjectIdentifier(stateType))
